requirements.txt
moviepy==1.0.3
openai-whisper==20231117
pydub==0.25.1
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
tqdm==4.66.1
imageio==2.31.5
imageio-ffmpeg==0.4.9
Pillow==10.1.0
